Finishes updates to product view
