+++ src/Components/Theme.tsx
@@ -30,9 +30,9 @@ export const themeProps = {
     black10: "#E5E5E5",
     black4: "#F6F6F6",
     white: "#fff",
+    white100: "#fff",
     black: "#000",
     green: "#44524A",
-    black40: "",
     lightGreen: "#989F9B",
     lightGray: "#DFDFDF",
     gray: "#7F7F7F",
